Split modify_files into multiple files
