+++ scripts/.gitignore
@@ -1,0 +1,2 @@
+*.rst  
+*modified.*
