+++ scripts/word.py
@@ -1,0 +1,10 @@
+"""
+Interacting with MS Word
+------------------------
+
+Requires ``python-docx`` e.g. ``pip3 install docx``
+
+
+"""
+
+import docx
